+++ routes/product.js
@@ -4,6 +4,13 @@ const Product = require("../models/product");
 
 //create product
 router.post('/', verifyTokenAndAmin, async (req, res)=>{
+
+    //validate request
+    if(!req.body.title) return res.status(400).send({message : "Ooops. Product title is required.."})
+    if(!req.body.description) return res.status(400).send({message : "Ooops. Product description is required.."})
+    if(!req.body.img) return res.status(400).send({message : "Ooops. Product image link is required.."})
+    if(!req.body.price) return res.status(400).send({message : "Ooops. Product price is required.."})
+
     const newProduct = new Product(req.body);
 
     try{
@@ -16,6 +23,10 @@ router.post('/', verifyTokenAndAmin, async (req, res)=>{
 
 //update product
 router.put('/:id', verifyTokenAndAmin, async (req, res)=>{
+
+    //validate request
+    if(!req.params.id) return res.status(400).send({message : "Ooops. Product id is required.."})
+
     try{
         const updatedProduct = await Product.findByIdAndUpdate(req.params.id, {
             $set: req.body},
@@ -28,6 +39,10 @@ router.put('/:id', verifyTokenAndAmin, async (req, res)=>{
 
 //delete product
 router.delete("/:id", verifyTokenAndAuthorization, async (req, res)=>{
+    
+    //validate request
+    if(!req.params.id) return res.status(400).send({message : "Ooops. Product id is required.."})
+
     try{
         await Product.findByIdAndDelete(req.params.id)
         res.status(200).json("Product has been deleted")
@@ -38,6 +53,10 @@ router.delete("/:id", verifyTokenAndAuthorization, async (req, res)=>{
 
 //get single product
 router.get('/findOne/:id', async (req, res)=>{
+    
+    //validate request
+    if(!req.params.id) return res.status(400).send({message : "Ooops. Product id is required.."})
+    
     try{
         const product = await Product.findById(req.params.id)
         res.status(200).json(product)
